[Backport release-3_40] Fix a bunch of coverity scan errors

+++ src/gui/qgslayermetadataresultsmodel.cpp
@@ -186,7 +186,7 @@ void QgsLayerMetadataResultsModel::reloadAsync()
 
 void QgsLayerMetadataResultsModel::resultsReady( const QgsLayerMetadataSearchResults &results )
 {
-  mFeedback->setProgress( mFeedback->progress() + 100 / QgsApplication::instance()->layerMetadataProviderRegistry()->layerMetadataProviders().count() );
+  mFeedback->setProgress( mFeedback->progress() + static_cast< double >( 100 ) / QgsApplication::instance()->layerMetadataProviderRegistry()->layerMetadataProviders().count() );
   beginInsertRows( QModelIndex(), mResult.metadata().count(), mResult.metadata().count() + results.metadata().count() - 1 );
   const QList<QgsLayerMetadataProviderResult> metadata { results.metadata() };
   for ( const QgsLayerMetadataProviderResult &result : std::as_const( metadata ) )
